requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.3.2
matplotlib==3.8.3
seaborn==0.13.2
azure-storage-blob==12.19.1
pyarrow==15.0.2
